Refactor the request and requestgroups detail pages to use the same component, which saves from having to download the same requestgroup twice when there is only one request, and gets rids of lots of duplication
